Test.Tasty.DejaFu

Description

This module allows using Deja Fu predicates with Tasty to test the behaviour of concurrent systems.

Testing

testAuto

Arguments

:: (Eq a, Show a)

=> (forall t. Conc t a)
The computation to test

-> TestTree

Automatically test a computation. In particular, look for deadlocks, uncaught exceptions, and multiple return values.

This uses the Conc monad for testing, which is an instance of MonadConc. If you need to test something which also uses MonadIO, use testAutoIO.

testDejafu

Arguments

:: Show a

=> (forall t. Conc t a)
The computation to test

-> TestName
The name of the test.

-> Predicate a
The predicate to check

-> TestTree

Check that a predicate holds.

testDejafus

Arguments

:: Show a

=> (forall t. Conc t a)
The computation to test

-> [(TestName, Predicate a)]
The list of predicates (with names) to check

-> TestTree

Variant of testDejafu which takes a collection of predicates to test. This will share work between the predicates, rather than running the concurrent computation many times for each predicate.

testAutoIO :: (Eq a, Show a) => (forall t. ConcIO t a) -> TestTree

Variant of testAuto for computations which do IO.

testDejafuIO :: Show a => (forall t. ConcIO t a) -> TestName -> Predicate a -> TestTree

Variant of testDejafu for computations which do IO.

testDejafusIO :: Show a => (forall t. ConcIO t a) -> [(TestName, Predicate a)] -> TestTree

Variant of testDejafus for computations which do IO.

Testing under Relaxed Memory

data MemType :: *

The memory model to use for non-synchronised CRef operations.

Constructors

SequentialConsistency
The most intuitive model: a program behaves as a simple interleaving of the actions in different threads. When a CRef is written to, that write is immediately visible to all threads.

TotalStoreOrder
Each thread has a write buffer. A thread sees its writes immediately, but other threads will only see writes when they are committed, which may happen later. Writes are committed in the same order that they are created.

PartialStoreOrder
Each CRef has a write buffer. A thread sees its writes immediately, but other threads will only see writes when they are committed, which may happen later. Writes to different CRefs are not necessarily committed in the same order that they are created.

testAuto'

Arguments

:: (Eq a, Show a)

=> MemType
The memory model to use for non-synchronised CRef operations.

-> (forall t. Conc t a)
The computation to test

-> TestTree

Variant of testAuto which tests a computation under a given memory model.

testAutoIO' :: (Eq a, Show a) => MemType -> (forall t. ConcIO t a) -> TestTree

Variant of testAuto' for computations which do IO.

testDejafu'

Arguments

:: Show a

=> MemType
The memory model to use for non-synchronised CRef operations.

-> Int
The maximum number of pre-emptions to allow in a single execution

-> Int
The maximum difference between the number of yield operations across all threads.

-> (forall t. Conc t a)
The computation to test

-> TestName
The name of the test.

-> Predicate a
The predicate to check

-> TestTree

Variant of testDejafu which takes a memory model and pre-emption bound.

testDejafus'

Arguments

:: Show a

=> MemType
The memory model to use for non-synchronised CRef operations.

-> Int
The maximum number of pre-emptions to allow in a single execution

-> Int
The maximum difference between the number of yield operations across all threads.

-> (forall t. Conc t a)
The computation to test

-> [(TestName, Predicate a)]
The list of predicates (with names) to check

-> TestTree

Variant of testDejafus which takes a memory model and pre-emption bound.

testDejafuIO' :: Show a => MemType -> Int -> Int -> (forall t. ConcIO t a) -> TestName -> Predicate a -> TestTree

Variant of testDejafu' for computations which do IO.

testDejafusIO' :: Show a => MemType -> Int -> Int -> (forall t. ConcIO t a) -> [(TestName, Predicate a)] -> TestTree

Variant of dejafus' for computations which do IO.